[SALES] 매출 업로드 Entity·Repository 최신 ERD 정합화

+++ src/main/java/com/memme/service/sales/upload/SalesUploadQueryService.java
@@ -169,7 +169,7 @@ public class SalesUploadQueryService {
         };
     }
 
-    private int zeroIfNull(Integer value) {
+    private long zeroIfNull(Long value) {
         return value == null ? 0 : value;
     }
